Separating out the logic for identification of individuals and the frame processing so that real-time analysis is easier to implement
Also added a new button for loading in new videos and images from the same directory, incase it is updated at runtime, the user now has that option instead of closing and relaunching the application. The pre-recorded analysis button is now disabled until resources (images and videos) are loaded, at which time it is enabled again.