Add multi-threading for pre-recorded video processing
Added custom function for generating chunks from an entire video, the video frames are read and saved to an array, that array is then passed into a function that splits it up into equal amounts and returns an array of chunks, which are made up of a number of frames. Each chunk is then processed on a separate thread, need to compute timings before keeping the functionality.